### 2025-02-21 e4f6f3e
Enhancement
Added WriteClean options to clean out old directories
