### 2025-02-21 9597649
Enhancement
Added WriteClean option to empty out old directories
