### 2025-03-17 #7821 @Kevin-Mattheus-Moerman
- [ ] Please study the above reference check :point_up: and see if you can address any of the reported potential DOI issues. Perhaps also check if all .bib entries are cited or if you can shorten it. You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again.
