### 2025-03-19 732ae3c
Documentation: added missing DOIs to bibtex for SPM25 paper
