+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={4},
   pages={189--210},
   year={1994},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={https://doi.org/10.1002/hbm.460020402}
 }
 
 @article{worsley1996unified,
@@ -61,7 +62,8 @@
   number={4},
   pages={1273--1302},
   year={2003},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/S1053-8119(03)00202-7}
 }
 
 @article{phillips2005empirical,
@@ -72,7 +74,8 @@
   number={4},
   pages={997--1011},
   year={2005},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.neuroimage.2004.10.030}
 }
 
 @inproceedings{brudfors2020flexible,
@@ -81,7 +84,8 @@
   booktitle={Medical Image Computing and Computer Assisted Intervention--MICCAI 2020: 23rd International Conference, Lima, Peru, October 4--8, 2020, Proceedings, Part III 23},
   pages={253--263},
   year={2020},
-  organization={Springer}
+  organization={Springer},
+  doi={https://doi.org/10.1007/978-3-030-59716-0\{_}25}
 }
 
 @article{yan2022factorisation,
@@ -91,7 +95,8 @@
   volume={15},
   pages={818604},
   year={2022},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
+  doi={https://doi.org/10.3389/fnins.2021.818604}
 }
 
 @article{andersson2003correct,
@@ -102,7 +107,8 @@
   number={2},
   pages={870--888},
   year={2003},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/s1053-8119(03)00336-7}
 }
 
 @article{friston2016bayesian,
@@ -112,7 +118,8 @@
   volume={128},
   pages={413--431},
   year={2016},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.neuroimage.2015.11.015}
 }
 
 @article{friston2018bayesian,
@@ -147,7 +154,8 @@
   volume={36},
   pages={390--405},
   year={2015},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.concog.2014.12.003}
 }
 
 @article{friston2017graphical,
@@ -158,7 +166,8 @@
   number={4},
   pages={381--414},
   year={2017},
-  publisher={MIT Press One Rogers Street, Cambridge, MA 02142-1209}
+  publisher={MIT Press One Rogers Street, Cambridge, MA 02142-1209},
+  doi={https://doi.org/10.1162/netn\{_}a\{_}00018}
 }
 
 @article{schwartenbeck2016computational,
@@ -168,7 +177,8 @@
   volume={3},
   number={4},
   year={2016},
-  publisher={Society for Neuroscience}
+  publisher={Society for Neuroscience},
+  doi={https://doi.org/10.1523/eneuro.0049-16.2016}
 }
 
 @article{friston2021active,
@@ -178,7 +188,8 @@
   volume={399},
   pages={107998},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.heares.2020.107998}
 }
 
 @article{donoghue2020parameterizing,
@@ -206,7 +217,8 @@
   volume={246},
   pages={118789},
   year={2022},
-  publisher={Elsevier}
[…]
