### 2025-03-19 423094b
Documentation: attempting to fix issue with underscores in DOIs
