+++ paper/paper.bib
@@ -85,7 +85,7 @@
   pages={253--263},
   year={2020},
   organization={Springer},
-  doi={https://doi.org/10.1007/978-3-030-59716-0\{_}25}
+  doi={https://doi.org/10.1007/978-3-030-59716-0{\_}25}
 }
 
 @article{yan2022factorisation,
@@ -167,7 +167,7 @@
   pages={381--414},
   year={2017},
   publisher={MIT Press One Rogers Street, Cambridge, MA 02142-1209},
-  doi={https://doi.org/10.1162/netn\{_}a\{_}00018}
+  doi={https://doi.org/10.1162/netn{\_}a{\_}00018}
 }
 
 @article{schwartenbeck2016computational,
