### 2025-03-19 47e34bb
Documentation: removed DOI URL in bibtex
