+++ paper/paper.bib
@@ -7,7 +7,7 @@
   pages={189--210},
   year={1994},
   publisher={Wiley Online Library},
-  doi={https://doi.org/10.1002/hbm.460020402}
+  doi={10.1002/hbm.460020402}
 }
 
 @article{worsley1996unified,
@@ -63,7 +63,7 @@
   pages={1273--1302},
   year={2003},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/S1053-8119(03)00202-7}
+  doi={10.1016/S1053-8119(03)00202-7}
 }
 
 @article{phillips2005empirical,
@@ -75,7 +75,7 @@
   pages={997--1011},
   year={2005},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.neuroimage.2004.10.030}
+  doi={10.1016/j.neuroimage.2004.10.030}
 }
 
 @inproceedings{brudfors2020flexible,
@@ -85,7 +85,7 @@
   pages={253--263},
   year={2020},
   organization={Springer},
-  doi={https://doi.org/10.1007/978-3-030-59716-0{\_}25}
+  doi={10.1007/978-3-030-59716-0{\_}25}
 }
 
 @article{yan2022factorisation,
@@ -96,7 +96,7 @@
   pages={818604},
   year={2022},
   publisher={Frontiers Media SA},
-  doi={https://doi.org/10.3389/fnins.2021.818604}
+  doi={10.3389/fnins.2021.818604}
 }
 
 @article{andersson2003correct,
@@ -108,7 +108,7 @@
   pages={870--888},
   year={2003},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/s1053-8119(03)00336-7}
+  doi={10.1016/s1053-8119(03)00336-7}
 }
 
 @article{friston2016bayesian,
@@ -119,7 +119,7 @@
   pages={413--431},
   year={2016},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.neuroimage.2015.11.015}
+  doi={10.1016/j.neuroimage.2015.11.015}
 }
 
 @article{friston2018bayesian,
@@ -155,7 +155,7 @@
   pages={390--405},
   year={2015},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.concog.2014.12.003}
+  doi={10.1016/j.concog.2014.12.003}
 }
 
 @article{friston2017graphical,
@@ -167,7 +167,7 @@
   pages={381--414},
   year={2017},
   publisher={MIT Press One Rogers Street, Cambridge, MA 02142-1209},
-  doi={https://doi.org/10.1162/netn{\_}a{\_}00018}
+  doi={10.1162/netn{\_}a{\_}00018}
 }
 
 @article{schwartenbeck2016computational,
@@ -178,7 +178,7 @@
   number={4},
   year={2016},
   publisher={Society for Neuroscience},
-  doi={https://doi.org/10.1523/eneuro.0049-16.2016}
+  doi={10.1523/eneuro.0049-16.2016}
 }
 
 @article{friston2021active,
@@ -189,7 +189,7 @@
   pages={107998},
   year={2021},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.heares.2020.107998}
+  doi={10.1016/j.heares.2020.107998}
 }
 
 @article{donoghue2020parameterizing,
@@ -218,7 +218,7 @@
   pages={118789},
   year={2022},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.neuroimage.2021.118789}
+  doi={10.1016/j.neuroimage.2021.118789}
 }
 
 @article{jafarian2019neurovascular,
[…]
