### 2025-03-19 2c2e1bb
Documentation: added more DOIs to bibtex
