+++ paper/paper.bib
@@ -18,7 +18,8 @@
   number={1},
   pages={58--73},
   year={1996},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/(SICI)1097-0193(1996)4:1<58::AID-HBM4>3.0.CO;2-O}
 }
 
 @article{friston1994analysis,
@@ -29,7 +30,8 @@
   number={2},
   pages={153--171},
   year={1994},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/hbm.460010207}
 }
 
 @article{josephs1997event,
@@ -40,7 +42,8 @@
   number={4},
   pages={243--248},
   year={1997},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/(SICI)1097-0193(1997)5:4<243::AID-HBM7>3.0.CO;2-3}
 }
 
 @article{ashburner2000voxel,
@@ -51,7 +54,8 @@
   number={6},
   pages={805--821},
   year={2000},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1006/nimg.2000.0582}
 }
 
 @article{friston2003dynamic,
@@ -126,7 +130,8 @@
   title={Bayesian model reduction},
   author={Friston, Karl and Parr, Thomas and Zeidman, Peter},
   journal={arXiv preprint arXiv:1805.07092},
-  year={2018}
+  year={2018},
+  doi={10.48550/arXiv.1805.07092}
 }
 
 @book{parr2022active,
@@ -144,7 +149,8 @@
   number={1},
   pages={1--49},
   year={2017},
-  publisher={MIT Press One Rogers Street, Cambridge, MA 02142-1209}
+  publisher={MIT Press One Rogers Street, Cambridge, MA 02142-1209},
+  doi={10.1162/NECO{\_}a{\_}00912}
 }
 
 @article{friston2015duet,
@@ -200,14 +206,16 @@
   number={12},
   pages={1655--1665},
   year={2020},
-  publisher={Nature Publishing Group US New York}
+  publisher={Nature Publishing Group US New York},
+  doi={10.1038/s41593-020-00744-x}
 }
 
 @article{medrano2024bsd,
   title={BSD: a Bayesian framework for parametric models of neural spectra},
   author={Medrano, Johan and Alexander, Nicholas A and Seymour, Robert A and Zeidman, Peter},
   journal={arXiv preprint arXiv:2410.20896},
-  year={2024}
+  year={2024},
+  doi={10.48550/arXiv.2410.20896}
 }
 
 @article{westner2022unified,
@@ -221,11 +229,15 @@
   doi={10.1016/j.neuroimage.2021.118789}
 }
 
-@article{jafarian2019neurovascular,
-  title={Neurovascular coupling: insights from multi-modal dynamic causal modelling of fMRI and MEG},
+@article{jafarian2020comparing,
+  title={Comparing dynamic causal models of neurovascular coupling with fMRI and EEG/MEG},
   author={Jafarian, Amirhossein and Litvak, Vladimir and Cagnan, Hayriye and Friston, Karl J and Zeidman, Peter},
-  journal={arXiv preprint arXiv:1903.07478},
-  year={2019}
+  journal={NeuroImage},
+  volume={216},
+  pages={116734},
+  year={2020},
+  publisher={Elsevier},
+  doi={10.1016/j.neuroimage.2020.116734}
 }
 
 @article{friston2019dynamic,
@@ -235,7 +247,8 @@
   volume={199},
   pages={730--744},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
[…]
