### 2025-03-19 81f3ffa
Documentation: minor bug fix to bibtex
