+++ paper/paper.bib
@@ -89,7 +89,7 @@
   pages={253--263},
   year={2020},
   organization={Springer},
-  doi={10.1007/978-3-030-59716-0{\_}25}
+  doi={10.1007/978-3-030-59716-0\_25}
 }
 
 @article{yan2022factorisation,
