### 2025-03-19 d5ee425
Documentation: Tidy-up of handling bibtex special characters
