+++ paper/paper.bib
@@ -150,7 +150,7 @@
   pages={1--49},
   year={2017},
   publisher={MIT Press One Rogers Street, Cambridge, MA 02142-1209},
-  doi={10.1162/NECO{\_}a{\_}00912}
+  doi={10.1162/NECO\_a\_00912}
 }
 
 @article{friston2015duet,
@@ -173,7 +173,7 @@
   pages={381--414},
   year={2017},
   publisher={MIT Press One Rogers Street, Cambridge, MA 02142-1209},
-  doi={10.1162/netn{\_}a{\_}00018}
+  doi={10.1162/netn\_a\_00018}
 }
 
 @article{schwartenbeck2016computational,
