### 2025-03-19 209b61f
Documentation: update to bibtex special characters
