### 2025-03-19 #7821 @pzeidman
Hi @Kevin-Mattheus-Moerman , I have updated the references but am having difficulty with three DOIs that contain underscores. I've tried to escape them using {\_} and \_ but without success. Any suggestions, please? (We are still working on the CONTRIBUTING.md file.) Thanks!
