### 2025-03-19 d4da489
Test: fix to test_regress_fmri_glm_dcm
