### 2025-03-20 #7821 @Kevin-Mattheus-Moerman
@xuanxu do these underscore containing DOIs need any special treatment?
