+++ paper/paper.bib
@@ -89,7 +89,7 @@
   pages={253--263},
   year={2020},
   organization={Springer},
-  doi={10.1007/978-3-030-59716-0{\_}25}
+  doi={10.1007/978-3-030-59716-0_25}
 }
 
 @article{yan2022factorisation,
@@ -150,7 +150,7 @@
   pages={1--49},
   year={2017},
   publisher={MIT Press One Rogers Street, Cambridge, MA 02142-1209},
-  doi={10.1162/NECO{\_}a{\_}00912}
+  doi={10.1162/NECO_a_00912}
 }
 
 @article{friston2015duet,
@@ -173,7 +173,7 @@
   pages={381--414},
   year={2017},
   publisher={MIT Press One Rogers Street, Cambridge, MA 02142-1209},
-  doi={10.1162/netn{\_}a{\_}00018}
+  doi={10.1162/netn_a_00018}
 }
 
 @article{schwartenbeck2016computational,
