### 2025-03-20 #7821 @pzeidman
Thanks @Kevin-Mattheus-Moerman and @xuanxu , sorry to have wasted your time! I've previously needed to escape underscores in bibtex, but you are correct it wasn't needed here. All parsing correctly now.
