### 2025-03-23 0a97858
Development work and minor upgrade of related routines for de novo learning (using a simple arcade game).
SVN r8493
