+++ toolbox/DEM/DEM_AtariII.m
@@ -32,7 +32,7 @@ rng(1)
 %==========================================================================
 Nr = 12;                                   % number of rows
 Nc = 9;                                    % number of columns
-Sc = 4;                                    % Spatial scaling
+Sc = 16;                                   % Spatial scaling
 
 % specify game
 %--------------------------------------------------------------------------
