### 2025-03-23 7a91496
Illustration of context sensitive preferences using DCM_law
SVN r8494
