### 2025-03-25 690d08e
Bug fix: Stopped segmentation from changing the boundary conditions used by spm_field
