### 2025-03-25 bbebb21
[Documentation] Draft contributor guidelines
