+++ CONTRIBUTING.md
@@ -1,0 +1,43 @@
+<!-- omit in toc -->
+# Contributing to SPM
+
+First off, thanks for taking the time to contribute! ❤️
+
+ Please see the Table of Contents below to skip to the relevant section about how we process contributions. It will make it a lot easier for us to incorporate your contribution. 
+
+## I Have a Question
+
+ If you want to ask a question, it is best to search for existing [Issues](https://github.com/spm/spm/issues) that might help you. You can respond to this issue to contribute to the discussion. There is also a historical [mailing list](http://jiscmail.ac.uk/cgi-bin/webadmin?A0=SPM) that has many years of contributions from the community. If you can't find any information about how to solve your problem please contribute to SPM so we can improve our software. 
+
+
+## Reporting Bugs and Issues
+
+We use GitHub issues to track bugs and errors. If you run into an issue with the project:
+
+- Make sure that you are using the latest version.
+- Read the relevant [documentation](https://www.fil.ion.ucl.ac.uk/spm/docs/).
+- Collect information about the issue:
+  - OS, Platform and Version (Windows, Linux, macOS, x86, ARM)
+- Open an [Issue](https://github.com/spm/spm/issues/new).
+	- Explain the behavior you would expect and the actual behavior.
+	- Provide us with code to reliably reproduce the issue.
+
+
+## Suggesting Enhancements
+
+This section guides you through submitting an enhancement suggestion for SPM, **including new features and minor improvements to existing functionality**. Enhancement suggestions are tracked as [GitHub issues](https://github.com/spm/spm/issues) and following the guidelines below will help us best process your suggestions.
+
+- Make sure that you are using the latest version of SPM.
+- Perform a [search](https://github.com/spm/spm/issues) to see if the enhancement has already been suggested. 
+- If not, open an enhancement issue with clear and descriptive title to identify the suggestion.
+- Describe the current behavior and explain which behavior you expected to see instead and why. 
+-  Highlight why the feature will be useful to the majority of our users and not just a small subset. If you're just targeting a minority of users, consider writing an independent toolbox.
+
+
+
+## Making Pull requests and contributing code
+If you feel confident in fixing the problem yourself, and are external to the development team, then please create a [pull request](https://github.com/spm/spm/pulls). In this case please see our developer guidelines on setting up [git](https://www.fil.ion.ucl.ac.uk/spm/docs/development/git/), our [code guidelines](https://www.fil.ion.ucl.ac.uk/spm/docs/development/guidelines/) and our guide to  software [testing](https://www.fil.ion.ucl.ac.uk/spm/docs/development/testing/).
+
+
+## Improving The Documentation
+We keep the documentation for SPM separate to the repository. If you would like to add something to our documentation [website](https://www.fil.ion.ucl.ac.uk/spm/docs/) please raise an issue on the [documentation repository](https://github.com/spm/spm-docs). 
