### 2025-03-27 #7821 @pzeidman
Thanks @samhforbes , just to let you know we have now added the [contributing file](https://github.com/spm/spm/blob/main/CONTRIBUTING.md).

All the best
Peter
