### 2025-03-28 4368994
upgrade to COVID and Arcade demos
SVN r8495
