+++ toolbox/DEM/spm_MDP_arcade.m
@@ -1,16 +1,12 @@
-function [MDP,hid,cid,con,RGB] = spm_MDP_arcade(Nr,Nc,Nd,Na,Np)
+function [MDP,RGB] = spm_MDP_arcade(Nr,Nc,Nd,Nb)
 % Creates an MDP structure for a simple arcade game
-% FORMAT [MDP,hid,cid,con,RGB] = spm_MDP_arcade(Nr,Nc,Nd,Na,Np)
+% FORMAT [MDP,hid,cid,con,RGB] = spm_MDP_arcade(Nr,Nc,Nd,Nb)
 %--------------------------------------------------------------------------
 % Nr    - number of rows
 % Nc    - number of columns
 % Nd    - number of initial states [default: 1]
-% Na    - returns rewards, costs and action
-% Np    - number of distractor pixels
+% Np    - number of bombs
 %
-% hid   - Hidden states corresponding to hits
-% cid   - Hidden states corresponding to misses
-% con   - output modalities reporting control
 % RGB   - display structure
 %
 % outcomes(1) - paddle
@@ -24,10 +20,9 @@ function [MDP,hid,cid,con,RGB] = spm_MDP_arcade(Nr,Nc,Nd,Na,Np)
 % Copyright (C) 2022-2023 Wellcome Centre for Human Neuroimaging
 
 if nargin < 1, Nr  = 12; end
-if nargin < 2, Nc  = 9; end
-if nargin < 3, Nd  = 2; end
-if nargin < 4, Na  = 1; end
-if nargin < 5, Np  = 0; end
+if nargin < 2, Nc  = 9;  end
+if nargin < 3, Nd  = 2;  end
+if nargin < 4, Nb  = 2;  end
 
 % preliminaries
 %--------------------------------------------------------------------------
@@ -39,6 +34,14 @@ Ns(5) = 2;              % momentum of ball verical
 Ns(6) = 3;              % level of target
 Ns(7) = 4;              % level of bomb
 NS    = prod(Ns);       % number of states
+
+% bomb offsets
+%--------------------------------------------------------------------------
+bo{1} = [ 0,0];
+bo{2} = [+2,1];
+bo{3} = [-3,2];
+bo{4} = [-1,0];
+bo{5} = [-3,1];
 
 % likelihoods and priors
 %==========================================================================
@@ -78,6 +81,7 @@ for x = 1:Ns(1)          % location of paddle
 
                             % current state
                             %----------------------------------------------
+                            reset = false;
                             s     = sub2ind(Ns,x,i,j,p,q,h,b);
 
                             % likelihoods
@@ -91,7 +95,6 @@ for x = 1:Ns(1)          % location of paddle
                                 %------------------------------------------
                                 ij = spm_index([Nr,Nc],g);
 
-
                                 % likelihoods for target
                                 %------------------------------------------
                                 if ij(1) <= h
@@ -99,11 +102,16 @@ for x = 1:Ns(1)          % location of paddle
                                     A{g}(3,s) = true;
                                 end
 
-                                % likelihoods for bomb
+                                % likelihoods for bombs
                                 %------------------------------------------
-                                if ij(1) == (b + Nr - Ns(7)) && ij(2) == i0
-                                    A{g}(:,s) = false;
-                                    A{g}(4,s) = true;
+                                for n = 1:Nb
+                                    ib = (i0 + bo{n}(1));
+                                    jb = rem(b + bo{n}(2) - 1,Ns(7)) + 1;
+                                    jb = (jb + Nr - Ns(7));
+                                    if ij(1) == jb && ij(2) == ib
+                                        A{g}(:,s) = false;
+                                        A{g}(4,s) = true;
+                                    end
                                 end
 
[…]
