### 2025-03-28 bea2c39
upgrade to COVID and Arcade demos
SVN r8495
