+++ toolbox/DEM/DEM_COVID_UK4.m
@@ -33,6 +33,13 @@ clc
 spm_figure('GetWin','SI'); clf;
 cd('C:\Users\karl\Dropbox\Coronavirus\Dashboard')
 
+% set final date for retrospective analyses. Full analysis: '01-Dec-2023'
+%==========================================================================
+global DATE
+date = '01-Mar-2022';
+DATE = date;
+%--------------------------------------------------------------------------
+
 % Files to be updated by hand
 %--------------------------------------------------------------------------
 % url = 'https://www.ons.gov.uk/file?uri=/peoplepopulationandcommunity/healthandsocialcare/conditionsanddiseases/datasets/coronaviruscovid19antibodydatafortheuk/2021/20210414covid19infectionsurveydatasets.xlsx'
@@ -41,67 +48,67 @@ cd('C:\Users\karl\Dropbox\Coronavirus\Dashboard')
 % writetable(webread(url,options),'place.csv');
 
 
-%% web options
+%% web options : all deprecated in Dec 2023
 %--------------------------------------------------------------------------
-options = weboptions('ContentType','table');
-options.Timeout = 40;
+% options = weboptions('ContentType','table');
+% options.Timeout = 40;
+% 
+% % download data and write to CSV files
+% %--------------------------------------------------------------------------
+% url = 'https://api.coronavirus.data.gov.uk/v2/data?areaType=nation&areaCode=E92000001&metric=newCasesBySpecimenDate&format=csv';
+% writetable(webread(url,options),'cases.csv');
+% url = 'https://api.coronavirus.data.gov.uk/v2/data?areaType=nation&areaCode=E92000001&metric=newTestsByPublishDate&format=csv';
+% writetable(webread(url,options),'tests.csv');
+% url = 'https://api.coronavirus.data.gov.uk/v2/data?areaType=nation&areaCode=E92000001&metric=newAdmissions&format=csv';
+% writetable(webread(url,options),'admissions.csv');
+% url = 'https://api.coronavirus.data.gov.uk/v2/data?areaType=overview&metric=newOnsDeathsByRegistrationDate&format=csv';
+% writetable(webread(url,options),'certified.csv');
+% url = 'https://api.coronavirus.data.gov.uk/v2/data?areaType=nation&areaCode=E92000001&metric=uniqueCasePositivityBySpecimenDateRollingSum&format=csv';
+% writetable(webread(url,options),'positivity.csv');
+% url = 'https://api.coronavirus.data.gov.uk/v2/data?areaType=nation&areaCode=E92000001&metric=newLFDTestsBySpecimenDate&format=csv';
+% writetable(webread(url,options),'lateralft.csv');
+% url = 'https://api.coronavirus.data.gov.uk/v2/data?areaType=nation&areaCode=E92000001&metric=cumVaccinesGivenByPublishDate&format=csv';
+% writetable(webread(url,options),'vaccines.csv');
+% 
+% % get cases by age (England)
+% %----------------------------------------------------------------------
+% url   = 'https://api.coronavirus.data.gov.uk/v2/data?areaType=nation&areaCode=E92000001&metric=newCasesBySpecimenDateAgeDemographics&format=csv';
+% tab   = webread(url,options);
+% vname = tab.Properties.VariableNames;
[…]
