### 2025-04-01 #7821 @samhforbes
Hi @stebo85, @cmaumet, @david-hofmann would any of you be interested in reviewing this software for JOSS? We do open review based on [these criteria](https://joss.readthedocs.io/en/latest/review_criteria.html).
