### 2025-04-05 0ca9257
Upgrade to arcade game demo
SVN r8496
