+++ toolbox/DEM/DEM_AtariIV.m
@@ -56,19 +56,15 @@ function MDP = DEM_AtariIV
 %--------------------------------------------------------------------------
 rng(1)
 
-EIGEN     = true;                            % eigenreduction
-INDUCTION = true;                            % inductive inference
-
 % Get game: i.e., generative process (as a partially observed MDP)
 %==========================================================================
 Nr = 12;                                     % number of rows
 Nc = 9;                                      % number of columns
-Nb = 2;                                      % number of bombs
+Nb = 3;                                      % number of bombs
 Nd = 1;                                      % number of random positions
 Sc = 32;                                     % spatial scaling
 T  = 2;                                      % time scaling
 C  = 32;                                     % log cost
-L  = [16,0];                                 % path lengths
 
 % get game in MDP form
 %--------------------------------------------------------------------------
@@ -91,9 +87,8 @@ spm_get_miss = @(o,id) find(o(id.contraint,:) > 1);
 %==========================================================================
 spm_figure('GetWin','Gameplay'); clf
 
-GDP.tau = 1;                                  % smoothness of random paths
-GDP.T   = 10000;                              % training length
-PDP     = spm_MDP_generate(GDP);              % generate play
+GDP.T = 10000;                              % training length
+PDP   = spm_MDP_generate(GDP);              % generate play
 
 % illustrate sequence of random play
 %---------------------------------------------------------------------------
@@ -106,14 +101,13 @@ for t = 1:128
     drawnow
 end
 
-% structure learning: grouping operators (iA,iB,iC,...)
+%% Structure learning: dependencies (iA,iB,iC,...)
 %==========================================================================
 MDP = spm_faster_structure_learning(PDP.O,S,Sc,T);
 
 % forget parameters
 %--------------------------------------------------------------------------
-Nm  = numel(MDP);
-for n = 1:Nm
+for n = 1:numel(MDP)
     for g = 1:numel(MDP{n}.a)
         MDP{n}.a{g} = [];
     end
@@ -122,151 +116,24 @@ for n = 1:Nm
     end
 end
 
-% Parameter learning: basins of attraction of goal states
+
+%% Parameter learning: basins of attraction
 %==========================================================================
 spm_figure('GetWin','Discovery'); clf
-
-NT    = 100;                         % number of frames per game 
-NS    = [];                          % number of states  
-NU    = [];                          % number of paths 
-NA    = [];                          % number of absorbing states
-NO    = [];                          % number of orphan states
-NH    = [];                          % number of goal states
-NG    = [];                          % number of transient goal states
-NF    = [];                          % number of goal states in orbit
-GDP.T = NT;
-for i = 1:200
-
-    % Accumulate states under random play
-    %======================================================================
-    if i < 1
-
-        % reuse continuous training data
-        %------------------------------------------------------------------
-        t = (0:(NT + T)) + rem(i,100 - 1)*NT;
-        
-    else
-
-        % regenerate training data from initial conditions
-        %------------------------------------------------------------------
[…]
