### 2025-04-06 dbff877
paralellized sliding time window inversion
