### 2025-04-16 0b69f7b
[Bug fix] More robust fitting of spheroid for AMM
