### 2025-04-17 #7821 @samhforbes
Hi @StevenZhang0116 and @michaelberks, would either of you be interested in reviewing this software for JOSS?
