### 2025-04-20 e9c1ca7
Development work on de novo learning (simplifications of faster structure learning)
SVN r8497
