### 2025-04-20 39f83ee
Development work on de novo learning (simplifications of faster structure learning)
SVN r8497
