### 2025-04-22 394be5e
Upgrade to diabetes demo
SVN r8499
