+++ toolbox/DEM/DEM_Diabetes.m
@@ -48,8 +48,8 @@ function DCM = DEM_Diabetes
 % where having a low risk (i.e., normal) genotype reduces reproductive
 % fitness when, and only when food insecure. The replicator dynamics are
 % situated in transitions among age groups, from child to adults and then
-% old. Mortality entails moving to the gene pool that slowly equilibria
-% rates over time. This means that replicator dynamics are mediated by
+% old. Mortality entails moving to the gene pool that slowly equilibriates
+% over time. This means that replicator dynamics are mediated by
 % replication rates that themselves depend upon genetic-environment
 % interactions. The clinical phenotype inherits from the BMI phenotype, in
 % the sense that the probability of moving from healthy to diabetic is
@@ -112,6 +112,8 @@ end
 %--------------------------------------------------------------------------
 spm_diabetes_T(pE);
 
+return
+
 % generate data
 %==========================================================================
 [Y,X] = spm_diabetes_gen(pE,M,U);
@@ -158,7 +160,7 @@ FS     = @(Y)real(sqrt(spm_vec(Y)));
 
 % model specification
 %==========================================================================
-M.Nmax = 4;                    % maximum number of iterations
+M.Nmax = 2;                    % maximum number of iterations
 M.G    = @spm_diabetes_gen;    % generative function
 M.FS   = FS;                   % feature selection  (link function)
 M.pE   = pE;                   % prior expectations (parameters)
@@ -188,7 +190,7 @@ spm_diabetes_plot(Y,X,M)
 
 % show parameter recovery
 %--------------------------------------------------------------------------
-spm_figure('GetWin','Latent states'), clf
+spm_figure('GetWin','Parameter recovery'); clf
 
 Vp      = spm_unvec(diag(Cp),pE);
 qP.P{1} = Ep;
@@ -210,36 +212,36 @@ function [P,C,str] = spm_diabetes_priors(U)
 
 % expectations
 %--------------------------------------------------------------------------
-P.size      = 1;               % population size
-P.diabetes  = 1/128;           % risk of diabetes | not low BMI
-P.excess    = 1/4;             % risk of diabetes | low BMI
-P.famine    = 1/32;            % P(food insecure | food secure)
-P.recovery  = 1/8;             % P(food secure | food insecure)
-P.nutrition = 1/4;             % P(normalisation of BMI | food secure)
-P.obesity   = 1/8;             % P(obesity | food secure)
-P.mortality = 1/16;            % P(morbidity | food insecure)
-P.fecundity = 1/2;             % P(replication | insecure and Low risk) 
-P.drift     = 1/128;           % genetic exchange between gene pools
-P.prod      = 1/8;             % sensitivity of production of food insecurity
-P.severity  = ones(size(U.T)); % severity of successive famines
+P.size       = 1;               % population size
+P.diabetes   = 1/128;           % risk of diabetes | not low BMI
+P.excess     = 1/4;             % risk of diabetes | low BMI
+P.famine     = 1/32;            % P(food insecure | food secure)
+P.recovery   = 1/8;             % P(food secure | food insecure)
+P.nutrition  = 1/4;             % P(normalisation of BMI | food secure)
+P.obesity    = 1/8;             % P(obesity | food secure)
+P.mortality  = 1/64;            % P(morbidity | food secure)
+P.risk       = 1/8;             % excess mortality risk if not low BMI
+P.epigenetic = 1/2;             % gene-environment interaction 
+P.prod       = 1/8;             % sensitivity of production of food insecurity
+P.severity   = ones(size(U.T)); % severity of successive famines
 
 % Variances (on log transformed parameters
 %--------------------------------------------------------------------------
 V     = exp(-2);               % uninformative priors
 W     = exp(-4);               % informative priors (weak)
 
[…]
