### 2025-04-23 #7821 @samhforbes
Hi @dbirman and @arokem would either of you be interested in reviewing this for JOSS?
