### 2025-04-24 #8103 @editorialbot
[REVIEW]: SPM 25: open source neuroimaging analysis software
**Submitting author:** @pzeidman (<a href="http://orcid.org/0000-0003-3610-6619">Peter Zeidman</a>)
**Repository:** https://github.com/spm/spm
**Branch with paper.md** (empty if default branch): spm-release-paper
**Version:** 25.01.02
