### 2025-04-24 65aff39
Bug Fix: more accurate definition of spheroid axes
