### 2025-04-24 c09245c
Bug Fix: No longer needs a "docs" folder to exist
