### 2025-04-24 bf1055e
Documentation: Removed a few empty lines in the help text to avoid dots in the markdown
