+++ config/spm_cfg.m
@@ -127,20 +127,16 @@ spmjobs.tag    = 'spm';
 spmjobs.name   = 'SPM';
 spmjobs.help   = {
     '%* Statistical Parametric Mapping', ...
-    '', ...
    ['Statistical Parametric Mapping refers to the construction and ' ...
     'assessment of spatially extended statistical processes used to ' ...
     'test hypotheses about functional imaging data. These ideas have ' ...
     'been instantiated in software that is called SPM.'], ...
-    '', ...
    ['The SPM software package has been designed for the analysis of ' ...
     'brain imaging data sequences. The sequences can be a series of ' ...
     'images from different cohorts, or time-series from the same ' ...
     'subject.'], ...
-    '', ...
    ['The current release is designed for the analysis of fMRI, PET, ', ...
-    'SPECT, EEG and MEG.'], ...
-    ''};
+    'SPECT, EEG and MEG.']};
 if exist('spm_cfg_eeg','file') == 2
     eeg = { spm_cfg_eeg };
 else
