+++ config/spm_cfg_deface.m
@@ -26,7 +26,7 @@ job.name  = 'De-face Images';
 job.val   = {images};
 job.help  = {
     'Strip the face from images, so individuals are more difficult to identify from surface renderings.'
-    'De-faced images are prefixed by ``anon_``.'
+    'De-faced images are prefixed by ``anon_``. Please note that this face-stripping may not always be reliable, so you are urged to visually inspect the end results prior to making data available to others.'
     }';
 job.prog  = @spm_deface;
 job.vout  = @vout;
