### 2025-04-25 cf4cbca
Documentation: Fix some spellings that appear in the markdown
