+++ config/spm_cfg_cdir.m
@@ -26,7 +26,7 @@ cdir.val     = {directory };
 cdir.help    = {
                 'This module is deprecated and has been moved to BasicIO.'
                 'Jobs which are ready to run may continue using it, but the module inputs can not be changed via GUI. Please switch to the BasicIO module instead.'
-                'This facilty allows programming a directory change. Directories are selected in the right listbox.'
+                'This facility allows programming a directory change. Directories are selected in the right listbox.'
 }';
 cdir.prog    = @my_job_cd;
 cdir.hidden  = true;
