+++ toolbox/DARTEL/tbx_cfg_dartel.m
@@ -111,7 +111,7 @@ initial         = cfg_exbranch;
 initial.tag     = 'initial';
 initial.name    = 'Initial Import';
 initial.val     = {matnames odir bb vox image GM WM CSF };
-initial.help    = {'Images first need to be imported into a form that Dartel can work with. If the default segmentation is used (ie the Segment button), then this involves taking the results of the segmentation (``*_seg_sn.mat``)/* \cite{ashburner05} */, in order to have rigidly aligned tissue class images. Typically, there would be imported grey matter and white matter images, but CSF images can also be included. The subsequent Dartel alignment will then attempt to nonlinearly register these tissue class images together. If the new segmentation routine is used (from the toolbox), then this includes the option to generate "imported" tissue class images. This means that a seperate importing step is not needed for it.'};
+initial.help    = {'Images first need to be imported into a form that Dartel can work with. If the default segmentation is used (ie the Segment button), then this involves taking the results of the segmentation (``*_seg_sn.mat``)/* \cite{ashburner05} */, in order to have rigidly aligned tissue class images. Typically, there would be imported grey matter and white matter images, but CSF images can also be included. The subsequent Dartel alignment will then attempt to nonlinearly register these tissue class images together. If the new segmentation routine is used (from the toolbox), then this includes the option to generate "imported" tissue class images. This means that a separate importing step is not needed for it.'};
 initial.prog = @spm_dartel_import;
 initial.vout = @vout_initial_import;
 % ---------------------------------------------------------------------
@@ -886,7 +886,7 @@ vox.strtype  = 'e';
 vox.val      = {[NaN NaN NaN]};
 vox.help     = {[...
 'Specify the voxel sizes of the deformation field to be produced. ',...
-'Non-finite values will default to the voxel sizes of the template image',...
+'Non-finite values will default to the voxel sizes of the template image ',...
 'that was originally used to estimate the deformation.']};
 % ---------------------------------------------------------------------
 %
@@ -924,7 +924,7 @@ nrm.help  = {[...
 'generated from unsmoothed native-space fMRI/PET data can be spatially ',...
 'normalised for a 2nd level analysis.'],[...
 'Two "preserve" options are provided.  One of them should do the ',...
-'equavalent of generating smoothed "modulated" spatially normalised ',...
+'equivalent of generating smoothed "modulated" spatially normalised ',...
 'images.  The other does the equivalent of smoothing the modulated ',...
 'normalised fMRI/PET, and dividing by the smoothed Jacobian determinants.'],[...
 'Note that the behaviour of Dartel is different from that of more conventional ',...
