+++ config/spm_cfg_md.m
@@ -34,7 +34,7 @@ md.tag          = 'md';
 md.name         = 'Make Directory (Deprecated)';
 md.val          = {basedir name};
 md.help         = {
-    'This facilty allows programming a directory change. Directories are selected in the right listbox.'
+    'This facility allows programming a directory change. Directories are selected in the right listbox.'
     'This module is deprecated and has been moved to BasicIO.'
     'Jobs which are ready to run may continue using it, but the module inputs can not be changed via GUI. Please switch to the BasicIO module instead.'
 }';
