+++ config/spm_cfg_deformations.m
@@ -432,7 +432,7 @@ pushfo.help   = {[...
 'generated from unsmoothed native-space fMRI/PET data can be spatially ',...
 'normalised for a 2nd level analysis.'],[...
 'Two "preserve" options are provided.  One of them should do the ',...
-'equavalent of generating smoothed "modulated" spatially normalised ',...
+'equivalent of generating smoothed "modulated" spatially normalised ',...
 'images.  The other does the equivalent of smoothing the modulated ',...
 'normalised fMRI/PET, and dividing by the smoothed Jacobian determinants.']};
 
