+++ external/fieldtrip/ft_topoplotER.m
@@ -24,6 +24,8 @@ function [cfg] = ft_topoplotER(cfg, varargin)
 %   cfg.baseline           = 'yes','no' or [time1 time2] (default = 'no'), see FT_TIMELOCKBASELINE or FT_FREQBASELINE
 %   cfg.baselinetype       = 'absolute' or 'relative' (default = 'absolute')
 %   cfg.trials             = 'all' or a selection given as a 1xN vector (default = 'all')
+%   cfg.magscale           = number, scaling to apply to the MEG magnetometer channels prior to display
+%   cfg.gradscale          = number, scaling to apply to the MEG gradiometer channels prior to display
 %   cfg.colormap           = string, or Nx3 matrix, see FT_COLORMAP
 %   cfg.marker             = 'on', 'labels', 'numbers', 'off'
 %   cfg.markersymbol       = channel marker symbol (default = 'o')
@@ -72,33 +74,27 @@ function [cfg] = ft_topoplotER(cfg, varargin)
 %                            'layout' to place comment as specified for COMNT in layout
 %                            [x y] coordinates
 %   cfg.interactive        = Interactive plot 'yes' or 'no' (default = 'yes')
-%                            In a interactive plot you can select areas and produce a new
-%                            interactive plot when a selected area is clicked. Multiple areas
-%                            can be selected by holding down the SHIFT key.
-%   cfg.directionality     = '', 'inflow' or 'outflow' specifies for
-%                            connectivity measures whether the inflow into a
-%                            node, or the outflow from a node is plotted. The
-%                            (default) behavior of this option depends on the dimord
-%                            of the input data (see below).
+%                            In an interactive plot you can select areas and produce a new interactive plot when a
+%                            selected area is clicked. Multiple areas can be selected by holding down the SHIFT key.
+%   cfg.directionality     = '', 'inflow' or 'outflow' specifies for connectivity measures whether the inflow into a 
+%                            node, or the outflow from a node is plotted. The (default) behavior of this option depends
+%                            on the dimord of the input data (see below).
 %   cfg.layout             = specify the channel layout for plotting using one of the supported ways (see below).
 %   cfg.interpolatenan     = 'yes' or 'no', whether to interpolate over channels containing NaNs (default = 'yes')
-%   cfg.figure             = 'yes' or 'no', whether to open a new figure. You can also specify a figure handle from FIGURE, GCF or SUBPLOT. (default = 'yes')
+%   cfg.figure             = 'yes', 'no', or 'subplot',  whether to open a new figure. You can also specify a figure
+%                            handle from FIGURE, GCF or SUBPLOT. (default = 'yes'). With multiple data inputs, 'subplot'
+%                            will make subplots in a single figure.
 %
-% For the plotting of directional connectivity data the cfg.directionality option
-% determines what is plotted. The default value and the supported functionality
-% depend on the dimord of the input data. If the input data is of dimord
-% 'chan_chan_XXX', the value of directionality determines whether, given the
-% reference channel(s), the columns (inflow), or rows (outflow) are selected for
-% plotting. In this situation the default is 'inflow'. Note that for undirected
-% measures, inflow and outflow should give the same output. If the input data is of
-% dimord 'chancmb_XXX', the value of directionality determines whether the rows in
-% data.labelcmb are selected. With 'inflow' the rows are selected if the
-% refchannel(s) occur in the right column, with 'outflow' the rows are selected if
-% the refchannel(s) occur in the left column of the labelcmb-field. Default in this
-% case is '', which means that all rows are selected in which the refchannel(s)
-% occur. This is to robustly support linearly indexed undirected connectivity
[…]
