+++ toolbox/DEM/spm_SARS_plot.m
@@ -29,7 +29,8 @@ function spm_SARS_plot(Y,X,Z,U)
 % https://www.telegraph.co.uk/global-health/science-and-disease/huge-regional-differences-intensive-care-bed-numbers-threaten/
 %--------------------------------------------------------------------------
 global CHOLD, if isempty(CHOLD); CHOLD = 1; end
-global DATE, if isempty(DATE); DATE = date; end
+global GHOLD, if isempty(GHOLD); GHOLD = 0; end
+global DATE,  if isempty(DATE); DATE = date; end
 
 % plot prior responses
 %==========================================================================
@@ -46,7 +47,7 @@ if nargin < 2
     % number of age groups if not specified
     %----------------------------------------------------------------------
     nN      = 4;
-    pE      = spm_SARS_priors(nN,DATE);
+    pE      = spm_SARS_priors(nN);
     M.T     = datenum(DATE) - datenum('01-01-2020','dd-mm-yyyy');
     [Y,X,Z] = spm_SARS_gen(pE,M,U);
     spm_SARS_plot(Y,X,Z,U)
@@ -105,11 +106,13 @@ xlabel('time (weeks)'),ylabel('number per day')
 title('Rates (per day)','FontSize',16)
 axis square, box off, set(gca,'XLim',[0, t(end)])
 legend('off'), legend(p(1:nu),un), legend('boxoff')
+if GHOLD, hold on, end
 
 subplot(3,2,2), set(gca,'ColorOrderIndex',1);
 plot(t,cumsum(Y));
 xlabel('time (weeks)'),ylabel('number of cases'), set(gca,'XLim',[0, t(end)])
 title('Cumulative cases','FontSize',16), axis square, box off
+if GHOLD, hold on, end
 
 % marginal densities
 %--------------------------------------------------------------------------
@@ -132,7 +135,8 @@ for i = 1:numel(X)
         ylabel('percent')
         title(str.factors{i},'FontSize',12), set(gca,'XLim',[0, t(end)])
         box off, legend(str.factor{i}(j(1:2))), legend('boxoff'), box off
-        
+        if GHOLD, hold on, end
+
         k = k + 1;
         subplot(6,2,k), set(gca,'ColorOrderIndex',1);
         j(1:2) = [];
@@ -140,14 +144,16 @@ for i = 1:numel(X)
         ylabel('percent')
         title(str.factors{i},'FontSize',12), set(gca,'XLim',[0, t(end)])
         box off, legend(str.factor{i}(j)), legend('boxoff'), box off
-        
+        if GHOLD, hold on, end
+
     elseif i == 2
         
         plot(t,X{i}(:,j(1:4))*100)
         ylabel('percent')
         title(str.factors{i},'FontSize',12), set(gca,'XLim',[0, t(end)])
         box off, legend(str.factor{i}(j(1:4))), legend('boxoff'), box off
-        
+        if GHOLD, hold on, end
+
         k = k + 1;
         subplot(6,2,k), set(gca,'ColorOrderIndex',1);
         j(1:4) = [];
@@ -155,14 +161,16 @@ for i = 1:numel(X)
         ylabel('percent')
         title(str.factors{i},'FontSize',12), set(gca,'XLim',[0, t(end)])
         box off, legend(str.factor{i}(j)), legend('boxoff'), box off
-        
+        if GHOLD, hold on, end
+
     else  
         
         plot(t,X{i}(:,j)*100)
         ylabel('percent')
         title(str.factors{i},'FontSize',12), set(gca,'XLim',[0, t(end)])
[…]
