+++ toolbox/DEM/DEM_masks_Cam.m
@@ -1,0 +1,283 @@
+function DEM_masks_Cam
+% Bayesian model reduction for COVID models
+% FORMAT DEM_mask_Cam
+%
+% This subroutine applies Bayesian model reduction to a DCM for COVID,
+% asking whether mask wearing can be treated as fixed parameters by
+% reducing its prior variance to 0.
+%__________________________________________________________________________
+% Copyright (C) 2020 Wellcome Centre for Human Neuroimaging
+
+% Karl Friston
+% Copyright (C) 2008-2022 Wellcome Centre for Human Neuroimaging
+
+% setup
+%==========================================================================
+
+% get posterior estimates
+%--------------------------------------------------------------------------
+clear
+cd('C:\Users\karl\Dropbox\Coronavirus\Dashboard')
+DCM = load('DCM_UK_tmp.mat','DCM');
+DCM = DCM.DCM;
+
+% unpack model and posterior expectations
+%--------------------------------------------------------------------------
+M   = DCM.M;                                 % model (priors)
+Ep  = DCM.Ep;                                % posterior expectation
+Cp  = DCM.Cp;                                % posterior covariances
+pE  = DCM.M.pE;                              % prior expectation
+pC  = DCM.M.pC;                              % prior covariances
+S   = DCM.Y;                                 % smooth timeseries
+U   = DCM.U;                                 % indices of outputs
+A   = DCM.A;                                 % age cohort
+
+Vp  = spm_unvec(spm_vec(diag(Cp)),Ep);       % posterior variances
+
+% Bayesian model comparison
+%==========================================================================
+param = {'msk','tra'};
+
+qE    = Ep;
+qC    = Cp;
+L     = [];
+str   = {};
+nP    = numel(param);
+for i = 1:nP
+
+    % can we ignore this effect? (i.e., vanishing prior variance)
+    %------------------------------------------------------------------
+    rE    = pE;
+    rC    = pC;
+    for j = 1:numel(pE.(param{i}))
+        rE.(param{i})(j) = rE.(param{i})(j) - 2;
+        rC.(param{i})(j) = rC.(param{i})(j)/exp(8);
+        F  = spm_log_evidence(qE,qC,pE,pC,rE,rC);
+        L(end + 1)   = 0 - F;
+        str{end + 1} = param{i};
+    end
+end
+
+% plot and display results of Bayesian model comparison
+%--------------------------------------------------------------------------
+spm_figure('GetWin','BMR-MASKS'); clf;
+subplot(2,1,1)
+bar(L), hold on
+plot([0 numel(L)],[5,5],'-.')
+set(gca,'XTickLabel',str,'YLim',[-1024 1024])
+title('Bayesian model comparison','Fontsize',16)
+xlabel('parameter'),ylabel('log Bayes factor')
+axis square, box off
+
+subplot(2,2,3), hold off
+x     = linspace(exp(-8),8/100,64);
+for i = 1:numel(Ep.msk)
+    p = spm_Npdf(log(x),Ep.msk(i),Vp.msk(i));
+    plot(x*100,p), hold on
+end
+title('Efficacy of face covering','Fontsize',16)
+xlabel('efficacy: % reduction in tranmission'), ylabel('posterior probability')
+xlabel('DCM parameter'), ylabel('correlation')
+axis square, box off, legend({'children','young','adults','elderly'})
+
+subplot(2,2,4), hold off
+j     = spm_fieldindices(pE,'msk');
+Rp    = spm_cov2corr(Cp);
+for i = j(:)'
+    plot(Rp(:,i)), hold on
+end
+title('Posterior correlations','Fontsize',16)
[…]
