### 2025-04-30 db8b207
minor changes to COVID routines
SVN r8502
