### 2025-05-01 49e95f2
Enhancement: refinement of channel select and inspection
