+++ toolbox/MEEGtools/spm_opm_psd.m
@@ -11,13 +11,17 @@ function [po,freq,indices] = spm_opm_psd(S)
 %   S.units         - units of measurement                  - Default: 'fT'
 %   S.constant      - constant line to draw as reference    - Default: 15
 %   S.wind          - function handle for window            - Default: @hanning
-%	  S.selectbad		  - enable selection of badchannels			  - Default: 0
-
+%	S.selectbad		- highlights and enables selection of 
+%                     bad channels			                - Default: 0
+%
 %
 % Output:
 %   psd             - power spectral density
 %   f               - frequencies psd is sampled at
-%	  indices				      - selected channel index 
+%	indices			- selected channel index
+%						To get labels use:
+%							plotted_lab = chanlabels(S.D,S.channels);
+%							sel_lab = plotted_lab(sel);
 %__________________________________________________________________________
 
 % Tim Tierney
@@ -36,16 +40,13 @@ if ~isfield(S, 'plot'),          S.plot = 0; end
 if ~isfield(S, 'D'),             error('D is required'); end
 if ~isfield(S, 'trials'),        S.trials=0; end
 if ~isfield(S, 'wind'),          S.wind=@hanning; end
-if ~isfield(S, 'interact'),      S.interact =1; end
-if ~isfield(S, 'selectbad'),		 S.selectbad = 0; end
-
-indices = [];
+if ~isfield(S, 'selectbad'),     S.selectbad = 0; end
 
 %-channel Selection
 %--------------------------------------------------------------------------
-
+indices = [];
 labs = S.channels;
-regex = {};
+regex = cell(1,length(labs));
 for i = 1:length(labs)
     if isa(labs,'cell')
         regex{i} = ['regexp_(',labs{i},')'];
@@ -81,7 +82,7 @@ catch
   wind = ones(N,1);
 end
 
-Nf= ceil((N+1)/2);
+% Nf= ceil((N+1)/2);
 coFac= max(wind)/mean(wind);
 wind = repmat(wind,1,length(chans));
 
@@ -133,25 +134,24 @@ end
 %--------------------------------------------------------------------------
 
 po = psdx;
+% Identify outliers
 med = repmat(median(po,2),1,size(po,2));
 ma = repmat(median(abs(po-med),2)*1.48,1,size(po,2));
 Z = (po-med)./ma;
 bin = Z>abs(spm_invNcdf(.025/(size(po,2)*100),0,1));
-badlabs = labs(sum(bin)>0)';
 
 if(S.plot)
-  f= figure();
-		h = plot(freq,po,'LineWidth',2);
-		hold on
+    figure();
+
+	h = plot(freq,po,'LineWidth',2);
+	hold on
+	for i = 1:numel(h)
+		tag = [labs{i}, ', Index: ' num2str(indchannel(S.D,labs{i}))];
+		set(h(i), 'UserData', false, 'tag',tag);
+	end
+
+    set(gca,'yscale','log')
     
-		for i = 1:numel(h)
-			set(h(i), 'UserData', true);
-      tag = [labs{i}, ', Index: ' num2str(indchannel(S.D,labs{i}))];
-      set(h(i), 'tag', tag);      
-      set(h(i), 'Color', [h(i).Color(1:3), 1]);
-    end
-    
-    set(gca,'yscale','log')  
     xp2 =0:round(freq(end));
     yp2=ones(1,round(freq(end))+1)*S.constant;
     p2 =plot(xp2,yp2,'--k');
@@ -161,6 +161,7 @@ if(S.plot)
     xlabel('Frequency (Hz)')
[…]
