### 2025-05-02 50e946c
Bug Fix: interactivity works with channel subsets
