### 2025-05-02 856cd60
Bug Fix: scaled chanpos for compatibility with image methods
