+++ toolbox/MEEGtools/spm_opm_create.m
@@ -331,7 +331,10 @@ if positions
         scalp_mesh = gifti(D.inv{1}.mesh.tess_scalp);
         lay = spm_get_anatomical_layout(pos, grad.label, double(scalp_mesh.vertices), fid_struct, 0);
         pos2d = transpose(lay.pos);
-        
+		
+        % Scale to range = 1 for spm_eeg image methods
+        pos2d = pos2d/max([max(pos2d(1,:)) - min(pos2d(1,:)); max(pos2d(2,:)) - min(pos2d(2,:))]);
+
         [sel1, sel2] = spm_match_str(lower(D.chanlabels), lower(lay.label));
         D = coor2D(D, sel1, num2cell(pos2d(:, sel2)));
         D.lay = lay;
