### 2025-05-02 87d29cc
Bug Fix: Neuro1 channels file to match other datatypes
