### 2025-05-02 5bec933
Bug Fix: bf_view_glass accepts inputs from earlier bf wizard methods
