### 2025-05-20 ac0bd82
Update spm_opm_create / sampling frq
manual assignment of sampling frequency improved for neuro-1 data
