### 2025-05-23 c5d3fba
Properly escape backslash in regular expression.
