+++ spm_dicom_convert.m
@@ -1321,7 +1321,7 @@ if isfield(Header,'GE_ImageType')
         end
     end
 else
-    if isfield(Header,'ImageType') && ~isempty(regexp(Header.ImageType,'\P\'))
+    if isfield(Header,'ImageType') && ~isempty(regexp(Header.ImageType,'\\P\\'))
         ImTyp = '-Phase';
     end
 end
