### 2025-05-23 55ed04e
Work around Octave crash when error stack has no column number (bug present in Octave 10.1 and older)
