### 2025-05-27 e6d7f6f
Bug Fix: Writing an error pessage caused an error
