+++ config/spm_cfg_realignunwarp.m
@@ -584,8 +584,9 @@ jm.help    = {
     'intensity that ensue as a consequence of the distortions. Think of a ' ...
     'semi-transparent coloured rubber sheet that you hold against a white background. ' ...
     'If you stretch a part of the sheet (induce distortions) you will see the ' ...
-    'colour fading in that particular area.']
-             }';
+    'colour fading in that particular area.'], ...
+    ['If the SCOPE toolbox was used to estimate a VDM, then use "Static field only", ' ...
+    'otherwise it may be better to select "None".']};
 jm.labels  = {
               'None'
               'Static field only'
