### 2025-05-27 079d097
Refactor: Reference to T1.nii was removed because it is not used anymore anywhere in the processing
