### 2025-05-29 #8103 @samhforbes
Hi all, thought I'd just check in here.
@stebo85 if you are happy with the paper as it stands i'm happy for you to tick off the item so we can proceed.
@dbirman I can see functionality is still being reviewed. Are there any issues here? Please let @pzeidman know if there's anything to work on here.
