### 2025-05-30 90af880
Changes for Octave compatibility
Two changes following Volkmar's pull request:
- Made saving consistent using spm_get_defaults('mat.format') to choose the format, rather than spm_check_version('matlab','7') where appropriate.
- Changed checks for Octave from exist('OCTAVE_VERSION','builtin') to strcmp(spm_check_version,'octave') for consistency
- Added Octave checks where need to avoid performing a MATLAB version check when Octave is being used.
