+++ spm_contrasts.m
@@ -320,8 +320,12 @@ SPM.xCon = xCon;
 
 % Check if SPM has changed. Save only if it has.
 %--------------------------------------------------------------------------
-if spm_check_version('matlab','8.0') >= 0, my_isequaln = @isequaln;
-else my_isequaln = @isequalwithequalnans; end
+is_octave = strcmp(spm_check_version,'octave');
+if is_octave || spm_check_version('matlab','8.0') >= 0
+    my_isequaln = @isequaln;
+else
+    my_isequaln = @isequalwithequalnans; 
+end
 if ~my_isequaln(tmpSPM,SPM)
     fprintf('\t%-32s: %30s','Saving SPM.mat','...writing');             %-#
     fmt = spm_get_defaults('mat.format');
