+++ compat/spm_adjmean_fmri_ui.m
@@ -595,14 +595,9 @@ Des = struct(...
         'aPMap',    [],...
         'EXnames',  EXnames,...
         'iX',       iX      );
-if spm_check_version('matlab','7') >= 0
-    save('SPMadj.mat','-V6',...
-        'SPMid','Des','V','c','cNames','W','Fnames','HPFc','HPF','sHPF','rX','nrX','rGX','GX');
-else
-    save('SPMadj.mat',...
-        'SPMid','Des','V','c','cNames','W','Fnames','HPFc','HPF','sHPF','rX','nrX','rGX','GX');
-end;
-
+save('SPMadj.mat',...
+        'SPMid','Des','V','c','cNames','W','Fnames','HPFc','HPF','sHPF','rX','nrX','rGX','GX',...
+        spm_get_defaults('mat.format'));
 
 %=======================================================================
 % - E N D
