### 2025-06-02 bf75bb5
Bug fix: Needs '\\P\\', rather than '\P\' to identify parts of ImageHeader containing '\P\'
