+++ toolbox/MEEGtools/spm_opm_create.m
@@ -170,7 +170,7 @@ end
 
 %- Position File check
 %----------------------------------------------------------------------
-try % to load a channels file
+try % to load a positions file
     posOri = spm_load(S.positions);
     positions =1;
 catch
@@ -276,9 +276,9 @@ if positions
     ori = [posOri.Ox,posOri.Oy,posOri.Oz];
     cl = posOri.name;
     
-    [sel1 sel2] = match_str(cl,channels.name);
+    [~, sel2] = match_str(cl,channels.name);
     
-    grad= [];
+    grad = [];
     grad.label = cl;
     grad.coilpos = pos;
     grad.coilori = ori;
@@ -649,11 +649,13 @@ function Snew = read_neuro1_data(Sold)
     for lin = (args.headerlength-4):args.headerlength
         line = fgetl(fid);
     end
-    channels = textscan(line, '%s', 'Delimiter', '\t');
+    fclose(fid);
+	
+	channels = textscan(line, '%s', 'Delimiter', '\t');
     channels = channels{1}(2:end);  % Cut out time variable
     channels(startsWith(channels, 'Comment')) = [];
     channels(startsWith(channels, 'Untitled')) = {'DataLogger'};
-    fclose(fid);
+    
     
     data = data(:,1:length(channels)); % remove comment channel
 
@@ -661,98 +663,6 @@ function Snew = read_neuro1_data(Sold)
     meg_chans = contains(units, 'pT');
     data(:,meg_chans) = data(:,meg_chans)*1e3;
     units(meg_chans) = {'fT'};
-
-    % Rename MEG channels to format "1-PX-X" (channel-opmname-channelaxis)
-    % if positions file provided
-    if isfield(Sold, 'chan2sens')
-        chan2sens = spm_load(Sold.chan2sens);
-
-        % First check format looks right
-        if ~isfield(chan2sens, 'channel')
-            warning('chan2sens.csv file must contain column named "channel".');
-            error('');
-        end
-        if ~isfield(chan2sens, 'sensor')
-            warning('chan2sens.csv file must contain column named "sensor".')
-            error('');
-        end
-
-        % Rename board slots from A1 to 1 (if there are any)
-        if isa(chan2sens.channel, 'cell')
-            board_slots = cell(length(chan2sens.channel), 1);
-            for slot = 1:length(chan2sens.channel)
-                if isa(chan2sens.channel{slot}, 'char')
-                    board_slots{slot} = regexp(chan2sens.channel{slot}, '[A-H]', 'match');
-                    if isa(board_slots{slot}, 'cell')
-                        board_slots{slot} = board_slots{slot}{1};
-                    end
-                end
-            end
-            rename_inds = cellfun(@isempty, board_slots);
-            rename_inds = ~rename_inds;
-            board_numbers = zeros(size(board_slots));
-            board_numbers(rename_inds) = cellfun(@(x)double(upper(x))-double('A')+1, board_slots(rename_inds));
-            chans = zeros(size(board_numbers));
-            chans(rename_inds) = (board_numbers(rename_inds)-1)*8 + cellfun(@(x)str2double(x(2)), chan2sens.channel(rename_inds));
-            chans(~rename_inds) = cellfun(@(x)str2double(x), chan2sens.channel(~rename_inds));
-            chan2sens.channel = chans;
-            clear chans
-        end
-
-        % Rename channels
-
-        % Handle different naming systems in data
-        if ~any(contains(channels(meg_chans), '_'))
-        % If the .lvm channels are labeled 'X9' etc.
-            meg_channos = cellfun(@(x)str2double(x(2:end)), channels(meg_chans));
[…]
