### 2025-06-16 d208523
Feature: First attempt at tools for computing cortical depth
