### 2025-06-16 5fb2456
Bug fix: Added the topology.mat file (and increased the iterations for computing distances)
