### 2025-06-18 b845395
Bug fix: Writing tissue maps from the nonparametric segmentation now works
