### 2025-06-20 3b3e5a7
Enhancement: use more liberal option format in realignunwarp to enable static correction without dynamic field modelling
