### 2025-06-20 #8103 @samhforbes
Hi @pzeidman the only thing I can see in the paper to correct is that Westner et al the journal should be rendered as Neuroimage rather than NeuroImage for consistent capitalisation with Friston et al 2016.
