### 2025-06-22 3e4d234
Documentation: Minor corrections to paper bibliography
