+++ paper/paper.bib
@@ -49,7 +49,7 @@
 @article{ashburner2000voxel,
   title={Voxel-based morphometry—the methods},
   author={Ashburner, John and Friston, Karl J},
-  journal={Neuroimage},
+  journal={NeuroImage},
   volume={11},
   number={6},
   pages={805--821},
@@ -61,7 +61,7 @@
 @article{friston2003dynamic,
   title={Dynamic causal modelling},
   author={Friston, Karl J and Harrison, Lee and Penny, Will},
-  journal={Neuroimage},
+  journal={NeuroImage},
   volume={19},
   number={4},
   pages={1273--1302},
@@ -106,7 +106,7 @@
 @article{andersson2003correct,
   title={How to correct susceptibility distortions in spin-echo echo-planar images: application to diffusion tensor imaging},
   author={Andersson, Jesper LR and Skare, Stefan and Ashburner, John},
-  journal={Neuroimage},
+  journal={NeuroImage},
   volume={20},
   number={2},
   pages={870--888},
@@ -118,7 +118,7 @@
 @article{friston2016bayesian,
   title={Bayesian model reduction and empirical Bayes for group (DCM) studies},
   author={Friston, Karl J and Litvak, Vladimir and Oswal, Ashwini and Razi, Adeel and Stephan, Klaas E and Van Wijk, Bernadette CM and Ziegler, Gabriel and Zeidman, Peter},
-  journal={Neuroimage},
+  journal={NeuroImage},
   volume={128},
   pages={413--431},
   year={2016},
@@ -243,7 +243,7 @@
 @article{friston2019dynamic,
   title={Dynamic causal modelling revisited},
   author={Friston, Karl J and Preller, Katrin H and Mathys, Chris and Cagnan, Hayriye and Heinzle, Jakob and Razi, Adeel and Zeidman, Peter},
-  journal={Neuroimage},
+  journal={NeuroImage},
   volume={199},
   pages={730--744},
   year={2019},
