### 2025-06-22 d8029c4
Documentation: added ORCID ids to paper.md
