### 2025-06-22 #8103 @pzeidman
Thanks @samhforbes . I have done the following:

- Updated all references to consistently say NeuroImage (the correct capitalization from the journal)
- Added authors' ORCID IDs where available
- Generated a release of the paper branch and uploaded it to Figshare:

URL: https://figshare.com/articles/dataset/SPM_25_01_JOSS_paper_source_code_/29377211?file=55548755
DOI: 10.6084/m9.figshare.29377211

Many thanks
Peter
