### 2025-06-23 #8103 @samhforbes
Hi @pzeidman the version is currently 25.01.02. Are you happy with that?
