+++ toolbox/DEM/DEM_swarm.m
@@ -1,44 +1,46 @@
-function DEM = DEM_morphogenesis
-% This routine illustrates self-assembly or morphogenesis under active
+function DEM = DEM_swarm
+% This routine illustrates self-organisation and swarming under active
 % inference (free energy minimisation).  It exploits the fact that one can
-% express a systems (marginal) Lyapunov function in terms of a variational
-% free energy.  This means that one can prescribe an attracting set in
-% terms of the generative model that defines variational free energy.  In
-% this example, the attracting set is a point attractor in the phase space
-% of a multi-celled organism: where the states correspond to the location
-% and (chemotactic) signal expression of 16 cells.  The generative model
-% and process are remarkably simple; however, the ensuing migration and
-% differentiation of the 16 cells illustrates self-assembly - in the sense
-% that each cell starts of in the same location and releasing the same
-% signals.  In essence, the systems dynamics rest upon each cell inferring
-% its unique identity (in relation to all others) and behaving in accord
-% with those inferences; in other words, inferring its place in the
-% assembly and behaving accordingly.  Note that in this example there are
-% no hidden states and everything is expressed in terms of hidden causes
-% (because the attracting set is a point attractor)  Graphics are produced
-% illustrating the morphogenesis using colour codes to indicate the cell
-% type - that is interpreted in terms of genetic and epigenetic
-% processing.
+% express a joint-systems (marginal) Lyapunov function in terms of a
+% variational free energy. This means that one can prescribe an attracting
+% set in terms of the generative model that defines variational free
+% energy.  In this example, the attracting set is an orbit in the phase
+% space of an ensemble (e.g., swarm of drones): where the states correspond
+% to the location and (radio) signal expression of 8 agents. In essence,
+% the systems dynamics rest upon each agent inferring its unique identity
+% (in relation to all others) and behaving in accord with those inferences;
+% in other words, inferring its place in the assembly and realising those
+% inferences by moving.  In this example, the behaviour is rendered
+% autonomous by including hidden states that evolve in time. This allows
+% the frame of reference (within which agents communicate) to change
+% systematically. In this example, this frame of reference uses affine
+% transformations to simulate sequences of swarm dynamics.
+%
+% This scheme can be compared with the morphogenesis simulations. The key
+% difference is that we are now dealing with an autonomous system in which
+% behaviour is prescribed by the internal (autonomous) dynamics of the
+% generative model.
 % _________________________________________________________________________
  
 % Karl Friston
 % Copyright (C) 2008-2022 Wellcome Centre for Human Neuroimaging
  
- 
 % preliminaries
 %--------------------------------------------------------------------------
-clear global
 rng('default')
-SPLIT    = 0;                              % split: 1 = upper, 2 = lower
-N        = 32;                             % length of process (bins)
+N  = 256;                                 % length of process (time bins)
+
+% universal time
+%--------------------------------------------------------------------------
[…]
