### 2025-06-23 99bd6d0
Addition of Swarm demo, based upon pattern formation. Minor change to spm_ADEM to accommodate action structures.
SVN r8503
