### 2025-06-24 2258f80
Bug fix: Writing spatially normalised tissue maps now works better for thick-sliced scans
Known issues remaining:
1) Much slower to write these data
2) Uses functionalityt from MB, which is a toolbox
