### 2025-06-26 3f82ab1
Bug fix: Missing data (outside the field of view) handled better in the spatially normalised images
