+++ spm_preproc_write8.m
@@ -39,8 +39,6 @@ tpm = res.tpm;
 if ~isstruct(tpm) || ~isfield(tpm, 'bg1')
     tpm = spm_load_priors8(tpm);
 end
-d1      = size(tpm.dat{1});
-d1      = d1(1:3);
 M1      = tpm.M;
 
 % Define orientation and field of view of any "normalised" space
@@ -114,7 +112,7 @@ for n=1:N
         create(chan(n).Nc);
     end
 
-    if bf(n,1),
+    if bf(n,1)
         chan(n).Nf      = nifti;
         chan(n).Nf.dat  = file_array(fullfile(pth1,['BiasField_', nam1, '.nii']),...
                                      res.image(n).dim(1:3),...
@@ -149,7 +147,7 @@ if do_defs
         Ndef.descrip = 'Inverse Deformation';
         create(Ndef);
     end
-    if df(2) || any(any(tc(:,[2,3,4]))) || nargout>=1,
+    if df(2) || any(any(tc(:,[2,3,4]))) || nargout>=1
         y = zeros([res.image(1).dim(1:3),3],'single');
     end
 end
@@ -167,11 +165,11 @@ for z=1:length(x3)
         f          = spm_sample_vol(res.image(n),x1,x2,o*x3(z),0);
         bf         = exp(transf(chan(n).B1,chan(n).B2,chan(n).B3(z,:),chan(n).T));
         cr{n}      = bf.*f;
-        if ~isempty(chan(n).Nc),
+        if ~isempty(chan(n).Nc)
             % Write a plane of bias corrected data
             chan(n).Nc.dat(:,:,z,chan(n).ind(1),chan(n).ind(2)) = cr{n};
         end
-        if ~isempty(chan(n).Nf),
+        if ~isempty(chan(n).Nf)
             % Write a plane of bias field
             chan(n).Nf.dat(:,:,z,chan(n).ind(1),chan(n).ind(2)) = bf;
         end
@@ -335,7 +333,7 @@ end
 
 % Write tissues if necessary
 for k1=1:Kb
-    if tc(k1),
+    if tc(k1)
         Nt      = nifti;
         Nt.dat  = file_array(fullfile(pth,['c', num2str(k1), nam, '.nii']),...
                              d, [spm_type('uint8') spm_platform('bigend')],...
@@ -394,9 +392,7 @@ for k1=1:size(tc,1)
         if tc(k1,1)
             Ni      = nifti;
             Ni.dat  = file_array(fullfile(pth,['rc', num2str(k1), nam, '.nii']),...
-                                 odim,...
-                                 [spm_type('float32') spm_platform('bigend')],...
-                                 0,1,0);
+                                 odim, 'float32', 0, 1, 0);
             Ni.mat         = mat;
             Ni.mat_intent  = 'Aligned';
             Ni.mat0        = mat0;
@@ -419,53 +415,28 @@ wc  = cell(size(cls));
 mwc = cell(size(cls));
 if ~any(tc(:)), return; end
 
-if any(any(tc(:,[1 3])))
-    C = zeros([d1,Kb],'single');
-end
-
 volsc = abs(det(M0(1:3,1:3))/det(M1(1:3,1:3)));
+sd    = spm_mb_shape('samp_dens',M1,M0);
 spm_progress_bar('init',Kb,'Warped Tissue Classes','Classes completed');
 for k1 = 1:Kb
     if ~isempty(cls{k1})
         c = single(cls{k1})/255;
-        if any(any(tc(:,[1 3]))) | any(any(tc(k1,[2 4])))
-            sd    = spm_mb_shape('samp_dens',M1,M0);
[…]
